Switch to HMI instead of LCD and Buttons in LCD_Task
